P2. Added Displaying Score + Organized script.js
ScoreManager displays the score of each paddle and tracks if a player has used one of their points to fire a shot.
Uses a while loop.
Organized Script.js so things are grouped together and easier to find.
Everything else is basically a mess right now so I've commented out the other features so I can focus on one feature at a time.