P2. Added Background Visuals
This took some fucking time, originally I wanted to make the background in script using the while loop but I couldn't figure out how to have the stars display without reshuffling everything (i.e. static). Settled on making an array with 7 background images stored in it, and a random selector that picks from the list at the start of every game. Reshuffles after a restart.