Ex8. Modified Spacelove + Added Spacelove to game
Ooof this is a big commit, I worked within Spacelove to make the kind of game I wanted, lots of tweaking, removing controles for the ships and giving them some kind of auto pilot. Then I added space love to the game and changed the properties so they'd be placed where i want them. Still have loads of work to do but I am making progress, lots to add not enough time.